requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
boto3==1.28.2
botocore==1.31.2
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.3
django-cleanup==7.0.0
django-debug-toolbar==4.1.0
django-storages==1.13.2
flake8==6.0.0
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
kombu==5.3.1
mccabe==0.7.0
Pillow==9.5.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
redis==4.6.0
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.6.2
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
